add option for user to set manual breakpoint
Previously breakpoint was hard-coded to `MAIN__` which works nicely for
FORTRAN programs but not for C/C++.

I have now added option `--breakpt` which allows user to manual set a
breakpoint at any valid function.

Also I have added logic to autoselect the "main" breakpoint as a default
for C and Fortran programs.