add uplot (youplot) library for plotting variables
uplot is a command line tool that draw plots on the terminal.

This helps to keep the user in terminal and not breakout to a GUI from
matplotlib. However, matplotlib is still available. And it could be
beneficial to use matplotlib if the number of debug ranks (-s) is
greater than e.g. 10.