feat: add input filter to interaction
The input filter can be used to arbitrarily hook commands given during
an `interact` session to do arbitrary things.

This commit implements hooks for `quit`, `q`, `exit`, and `Ctrl-d` to
avoid killing the gdb instance that is being interacted with, and
instead safely returns the user back to the mdb shell.