fix: dont reuse memory in filter closures
Move the list initializiation from the closure arguments to within the
body of the wrapper so that the memory is always reinitialized to a
known state before the closure is handed to `pexpect`.

To allow the same closure function to be reused over multiple interacts,
also ensure that the buffer is emptied when EOF is given.