fix: remove gdb shell input persistence
When exiting an `interact` with a GDB shell, whatever was currently
entered in the command string persists between `interact`. This can
lead to annoying and unwanted behaviour from unintended strings
prefixing the next command.

To avoid this, we send a single `Ctrl-C` character to GDB *after* the
interaction session has been closed to clear whatever was in the input
buffer.