Improve handling of closed debug processes
Currently the debug processes can be closed but the user may not be
notified. This can lead to them submitting commands that don't do
anything with no feedback.

These changes to `mdb/mdb_client.py` and `mdb/mdb_shell.py` provide a
warning when all debug processes `client.dbg_procs` are closed.

This may also paves the way for handling debug session restarts (pending
a re-write of the mdb launcher).