chore: refactor cert and key paths
Preferring a function that returns the paths instead of different hard
coded variables. This will later also grant flexibility if the paths
should be read e.g. from an env var or config file.

Also means the paths can be just modified in one place in order to
propagate the changes throughout the project.