feat: make the exchange server dumber
Previously the exchange server was dumb, but it still was smart enough
to know that every message from the client must result in a message back
from the debuggers. This modicum of intelligence meant it was deaf to
the client sending additional messages, such as interrupts, until the
debuggers replied to the first message.

Exchange server now sends whatever it receives from the debuggers to the
clients, and whatever it receives from the clients to the debuggers
without trying to verify the order.