feat: concurrent debuggers
Similar to the problems the exchange server was having, the debuggers
would not receive the interrupt command as they were still waiting for
the `pexpect` process to return something sensible to send back to the
exchange.

Modified so that each command launches a new task which tries to do the
previous behaviour, but allows the debugger to simultaneously listen for
new commands coming in. Also tracks the last run command so that
interrupts (or even status polling down the line) can interact with the
currently running task.