chore: use pyproject.toml
See PEP 518 and PEP 621. Distutils is also being discontinued so want to
always the modern tools.

The only thing that doesn't play nicely at the moment is flake8, which
wants its configuration in `tox.ini` instead of `pyproject.toml`. But
this is also pretty standard, and even used in the sample Python project
that PyPa maintain.