tests: print server logs and allow root
Added pipes to the stderr and stdout so we can see what the exchange
server is trying to do. This means also adding `--log-level=DEBUG` so we
get decent information if things fail.

I noticed then, whilst running the tests in a docker container, that the
output was complaining that the MPI was trying to run as root, which is
conventionally disallowed. Added the necessarily environment variables
to get past that.

As far as I can tell, the tests are now running correctly in the CI when
I test with `act`, but the outputs are failing due to some differences
in the output.