tests: add connection test
This test is basically implicit in running the different backend tests,
however it gives us one spectacular piece of information with a single
glance: did the backend test fail because something went wrong with mdb
or just because the output didn't match up?

Saves seconds, but does marvels for your sanity.