ci: remove coverage
Coverage bad???? When coverage is enabled on pytest, the test either
hangs or the client/server output nothing (and ergo test fails).  Maybe
coverage is doing something weird with child processes?

Running coverage here might not be the best idea anyway, as the
subprocess calls mean coverage wont actually see which parts of the
client/server code we are hitting.

It's going to be tricky to come up with a test harness that all runs
within the same Python process, but that is what would be needed for
coverage to make sense.