tests: directly call attach in timeout
This avoids some race-condition-like behaviour where sometimes the
timeout would repeat the connection attempts and other times it would
not, making it difficult to validate the output.

The thing we really care about in this test is that the right error
message is produced (which is only produced when it has attempted N many
connections), so can avoid having to run as a subprocess and can instead
call the attach command directly.