feat(mdb_attach): add attach_shell function
This is to address running the tests in a more reproducible and sensible
way. This new function returns the mdb shell that it has instantiated,
so that we can script behaviour in our tests without subprocess calls and
let Python handle errors, and coverage gather coverage.

The command-line attach method is now a guard for this new method,
validating the user's input and then closing the event loop.