tests: drive shell from test process
Instead of spawning a new process to run the shell in, recent commits
have made it possible to drive and control the mdb shell
programmatically. This allows us to test the shells behaviour from
within the test process, with will allow coverage to be re-enabled, and
for more granular output testing.

Some examples of this are included in this commit, however only a
minimal example of this is included.

Smoke tests are now also easier to define, as errors that occur in the
shell process are caught by the test runner and bubbled back to pytest
without us having to check for them manually.