feat(exchange): ensure debuggers with client
When the client connects, we now ensure there are debuggers connected
before we setup the message exchange. If the debuggers haven't yet
connected, we check periodically until a timeout interval expires.
If there are still no debuggers, we then send a `exchange_info` message
to the client, and kill the exchange server with an error.