fix(exchange): wait until debuggers init
Don't add the debuggers to the list of connected debuggers until _after_
the debuggers have told us they have initialised.

This is part of the reason the CI often failed: the docker containers
are constrained environments, so things don't necessarily happen in the
order you'd expect, so sometimes the client would start sending messages
after the debuggers had connected, but before they confirmed
initialisation. This then ends with the message ratchet being out of
sync, and everything deadlocks without any useful errors.

With this change (and some of the previous strict message checking), the
CI seems to pass consistently.