Revert "ci: remove coverage"
This reverts commit 6f1881d98b863b844deec2a2390d8c3e7ccecda4.

I think coverage was contributing to the overhead on the poor little
docker container that was already struggling to juggle three mdb
processes, two MPI processes, two gdb processes, the pytest, and the mdb
shell client.

This is likely why it seemed when coverage was enabled the CI was
failing more often, because the machine was _more_ over allocated and
therefore more likely to hit the race condition. With the messages made
stricter and some catches to ensure the debuggers are connected and
initialised correctly, disabling the coverage is no longer needed.