feat: mdb attach select is now set by exchange server
previously user could specify different ranks to those which are running
debuggers (specified by the `mdb launch command`). This could confuse
the user even though it wouldn't break anything.

Now we also check that the user picks ranks (via the `select` command)
that are available, according to the exchange server.

If user tries to select a rank outside of the available ranks they will
get a helpful error message.